More optimizations of the map rendering code. This time efficiency has been boosted by nearly 25%. I added three optimizations to get these gains.
First, I re-enabled lightmap texture tracking for non-GLSL surfaces *as long
as GLSL is enabled.* I've found that this does not result in any artefacts
like it does when used with GLSL disabled. I'm still not sure why that is.

Second, I added state tracking for tangentSpaceTransform data, internalizing
it so it can be compared efficiently between surfaces.

However, the biggest gain by far involved rearranging the VBO data. Now,
instead of the XYZ, ST, and LM data grouped together for each surface, the
entire VBO is partitioned into XYZ, ST, and LM sections. This means that the
vertex pointer and texcoord pointers need only be set *once,* when the VBO is
first bound. This results in a massive reduction in the number of GL calls.


git-svn-id: svn://svn.icculus.org/alienarena/trunk/source@3085 05fc5d3f-ce2a-0410-9b80-abac23765f80