Handle mesh scaling for shell effects in the vertex shader.
Just like before, the normal of each vertex is multiplied by a scaling factor
and added to the vertex coordinates for shell effects. Only now it's done in
GLSL. The useShell uniform value for the mesh vertex shader is now a float 
containing the scaling factor, instead of an integer with a boolean value.

This change unbreaks shell scaling for IQM models, and is a prerequisite for
GLSL lerping for MD2 models.


git-svn-id: svn://svn.icculus.org/alienarena/trunk/source@3813 05fc5d3f-ce2a-0410-9b80-abac23765f80