Massively improved mesh occlusion culling.
Seems to improve false negatives, allowing more meshes to be culled and giving
a performance boost on some maps. Also seems to fix a few false positives,
preventing some meshes from being culled that shouldn't be, like the cacti in
dm-wasteland.

The old version had false positives because it used a volume trace designed to
find any obstacle to the whole bounding box. This is great for physics, 
because a foot-high wall can block your movement, but not for occlusion
culling, because a foot-high wall won't block people from seeing you.

The old version also had false negatives because it used an axis-aligned
un-rotated bounding box which may not actually enclose the oriented/rotated
mesh. 

The new version now does nine point (non-volume) traces-- one at each vertex
of the *oriented* bounding box, and one at the mesh's origin. The idea is to
check if *any* part of the mesh is visible. False positives are still possible
(although I was unable to find any maps that had them,) but false negatives
are not anymore.


git-svn-id: svn://svn.icculus.org/alienarena/trunk/source@4335 05fc5d3f-ce2a-0410-9b80-abac23765f80