Mostly fix tangent issues in terrain meshes, fixing some rendering artifacts.
Artifacts fixed:
* some triangles lighting up at too great a distance from dynamic lights
* some triangles never lighting up at all from dynamic lights.

These were fixed by filtering out (for the purposes of tangent and normal
generation) any triangles with handedness 1 in terrain meshes.

In order to calculate handedness, the mesh loader was made to use the same
algorithm as the BSP loader to calculate tangents. This algorithm is now in
the function R_CalcTangent in r_math.c. 

If developer output is enabled, the mesh loader now warns if two triangles of
differing handedness use the same vertex. 

As part of this commit, handedness is now actually correct for mesh vertices,
instead of being hardcoded at 1. This is may fix other artifacts as well,
although I don't know for sure.



git-svn-id: svn://svn.icculus.org/alienarena/trunk/source@4623 05fc5d3f-ce2a-0410-9b80-abac23765f80