ref_gl: overhaul VBO batching, part 1.
The old design was to keep a statically allocated list of VBO batch
structures, which had the effect of limiting how many VBO batches could be
accumulated before they'd have to be flushed (i.e. rendered.) We'd lose
opportunities to merge two batches because the accumulator would need to be
flushed before the merge would happen.

In the new design, the VBO batching data structures are included in
msurface_t, so there's no limit to how many VBO batches can be accumulated. We
no longer miss out on opportunities to merge two batches. Furthermore, the VBO
batching system is inherently more flexible now because you no longer need to
check whether it has room for more surfaces; you need only be prepared to do
one render.

In the old design, the list of existing VBO batches had to be walked every
time a new surface was added, looking for an existing batch to add the surface
to or if none existed, inserting a new batch with the surface. This made the
act of adding a new surface into an O(n) operation.

In the new design, adding a new surface to the VBO batch accumulator is an
O(1) operation, because it just looks at whatever surfaces come before/after
the new surface within the VBO, and checks whether those neighboring surfaces
are part of existing batches.


git-svn-id: svn://svn.icculus.org/alienarena/trunk/source@4785 05fc5d3f-ce2a-0410-9b80-abac23765f80