ACEbot: no longer read/write "facts table" (item_table) to/from .nod files.
When it was read from the file, it was immediately re-generated anyway, so
there was no use in reading it at all.

Furthermore, the struct that was written contained a pointer, of all things,
so even if you tried to use the data it would probably cause segfaults. 

From now on, when a .nod file is read, the item table is skipped and ignored.

From now on, when a .nod file is written, the file header is given a
"num_items" value of zero, so even legacy executables won't try to read the
item table from the file. 

What this means is that the item_table_t structure can now be changed freely
(including changing its size) without breaking backwards compatibility with
old .nod files, and without breacking forwards compatibility of old game
executables with new .nod files.


git-svn-id: svn://svn.icculus.org/alienarena/trunk/source@4918 05fc5d3f-ce2a-0410-9b80-abac23765f80