renderer: more fixes to multiple lightstyles.
Mesh lighting level now updates every frame when needed to track lightstyles.
When no lightstyles are involved, it will still avoid updating when meshes
don't move.

Copy the pixel data for all lightstyles from lightmap files when applicable.
Before it was only copying the first lightstyle, even though qrad3 was doing
the right thing and putting them all in .lightamp files.

RecursiveLightPoint: handle lightmap files properly when working with multiple
lightstyles. This functionality had been broken because it was still assuming
the lightmap texture scale of 16 which is used for the BSP's built-in
lightmap.


git-svn-id: svn://svn.icculus.org/alienarena/trunk/source@5313 05fc5d3f-ce2a-0410-9b80-abac23765f80